speech_rec_chatbot_requirements.txt
action-tutorials-py==0.9.4
ament-copyright==0.9.7
ament-cppcheck==0.9.7
ament-cpplint==0.9.7
ament-flake8==0.9.7
ament-index-python==1.1.0
ament-lint==0.9.7
ament-lint-cmake==0.9.7
ament-package==0.9.5
ament-pep257==0.9.7
ament-uncrustify==0.9.7
ament-xmllint==0.9.7
apturl==0.5.2
argcomplete==2.0.0
arm-pack==0.0.0
atomicwrites==1.1.5
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
blis==0.2.4
Brlapi==0.7.0
catkin-pkg-modules==0.5.2
cbor==1.0.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
chatbot==0.0.0
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
Click==7.0
colcon-argcomplete==0.3.3
colcon-bash==0.4.2
colcon-cd==0.1.1
colcon-cmake==0.2.27
colcon-common-extensions==0.3.0
colcon-core==0.12.0
colcon-defaults==0.2.8
colcon-devtools==0.2.3
colcon-installed-package-information==0.1.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.15
colcon-output==0.2.12
colcon-override-check==0.0.1
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.2.4
colcon-pkg-config==0.1.0
colcon-powershell==0.3.7
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.1
colcon-ros==0.3.23
colcon-test-result==0.3.8
colcon-zsh==0.4.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cov-core==1.15.0
coverage==4.5.2
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.7
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
demo-nodes-py==0.9.4
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dlib==19.24.0
docutils==0.16
domain-coordinator==0.9.2
duplicity==0.8.12.0
empy==3.3.2
entrypoints==0.3
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
exceptiongroup==1.1.0
eye-movement==0.0.0
face-tracker==0.0.0
fasteners==0.14.1
flake8==3.7.9
flake8-blind-except==0.2.1
flake8-builtins==2.1.0
flake8-class-newline==1.6.0
flake8-comprehensions==3.10.1
flake8-deprecated==2.0.1
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-quotes==3.3.2
future==0.18.2
h5py==2.10.0
head-gestures==0.0.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
ifcfg==0.18
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==2.0.0
joblib==0.14.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
lark-parser==0.8.1
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mathparse==0.1.2
matplotlib==3.1.2
mccabe==0.6.1
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
murmurhash==1.0.9
netifaces==0.10.4
nltk==3.8.1
nose==1.3.7
nose2==0.9.1
notify2==0.3
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.7.0.68
osrf-pycommon==0.1.11
packaging==20.3
parameterized==0.7.0
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
Pint==0.20.1
pixel-ring==0.1.0
plac==0.9.6
pluggy==0.13.0
pocketsphinx==5.0.0
preshed==2.0.1
progressbar==2.5
protobuf==3.6.1
psutil==5.5.1
py==1.8.1
py-srvcli==0.0.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.13
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.21
pycups==1.9.73
pydocstyle==6.3.0
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
pygpu==0.7.6
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pytest==7.2.1
pytest-cov==2.8.1
pytest-repeat==0.9.1
pytest-rerunfailures==11.1
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyusb==1.2.1
pyxdg==0.26
PyYAML==5.1.2
quality-of-service-demo-py==0.9.4
regex==2023.3.23
reportlab==3.5.34
requests==2.28.2
requests-unixsocket==0.2.0
respeaker==0.6.2
roman==2.0.0
ros2action==0.9.12
ros2bag==0.3.9
ros2cli==0.9.12
ros2component==0.9.12
ros2controlcli==0.11.0
ros2doctor==0.9.12
ros2interface==0.9.12
ros2launch==0.11.7
ros2lifecycle==0.9.12
ros2multicast==0.9.12
ros2node==0.9.12
ros2param==0.9.12
ros2pkg==0.9.12
ros2run==0.9.12
ros2service==0.9.12
ros2topic==0.9.12
rosdep==0.22.1
rosdep-modules==0.22.1
rosdistro-modules==0.9.0
rosidl-runtime-py==0.9.1
rospkg==1.4.0
rospkg-modules==1.4.0
rpyutils==0.2.0
rqt==1.1.2
rqt-action==0.4.9
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-moveit==1.0.1
rqt-msg==1.0.5
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.5
rqt-robot-steering==1.0.0
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.2
rqt-top==1.0.2
rqt-topic==1.3.0
scikit-learn==0.22.2.post1
scipy==1.3.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
snowballstemmer==2.2.0
sos==4.4
sounddevice==0.4.5
soupsieve==1.9.5
spacy==2.1.9
speech-python==0.0.0
SpeechRecognition==3.9.0
spidev==3.6
SQLAlchemy==1.2.19
sros2==0.9.5
srsly==1.0.6
ssh-import-id==5.10
systemd-python==234
teleop-twist-keyboard==2.3.2
Theano==1.0.4
thinc==7.0.8
tomli==2.0.1
topic-monitor==0.9.4
tqdm==4.65.0
Twisted==18.9.0
txaio==2.10.0
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vcstool==0.3.0
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.1.8
webencodings==0.5.1
webrtcvad==2.0.10
wsaccel==0.6.2
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
